Share more code, optimizations
